[v2.4.x] prov/efa: Fix use-after-free in DC packet completion handling
DC packets had a race condition where RECEIPT acknowledgments could
arrive before all TX completions, causing premature txe release and
use-after-free errors when remaining TX completions accessed the
freed memory.

Fix by using efa_outstanding_tx_ops for completion tracking:
- DC packets use efa_outstanding_tx_ops to track TX completion status
- RECEIPT handler writes completion immediately (preserves DC semantics)
- txe release delayed until both TX completions and receipt received
- Added EFA_RDM_TXE_RECEIPT_RECEIVED flag to track receipt arrival
- Added efa_rdm_txe_dc_ready_for_release() with proper documentation

This prevents the race condition while maintaining existing DC
completion behavior for applications.

(cherry picked from commit 95bc897c1568c646cdc192ed78320adc86030f85)

+++ prov/efa/src/rdm/efa_rdm_ope.c
@@ -1963,7 +1963,7 @@ ssize_t efa_rdm_ope_repost_ope_queued_before_handshake(struct efa_rdm_ope *ope)
 	}
 }
 
-int efa_rdm_ope_process_queued_ope(struct efa_rdm_ope *ope, uint16_t flag)
+int efa_rdm_ope_process_queued_ope(struct efa_rdm_ope *ope, uint32_t flag)
 {
 	int ret = 0;
